requirements.txt
aiohttp>=3.9.0
aiohttp-socks>=0.8.0
aiosqlite>=0.19.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
backoff>=2.2.0
schedule>=1.2.0
